+++ client/src/rpc/impls/cfx.rs
@@ -11,8 +11,8 @@ use crate::rpc::{
         BlockHashOrEpochNumber, Bytes, CallRequest,
         CheckBalanceAgainstTransactionResponse, ConsensusGraphStates,
         EpochNumber, EstimateGasAndCollateralResponse, Filter as RpcFilter,
-        Log as RpcLog, Receipt as RpcReceipt, SendTxRequest,
-        SponsorInfo as RpcSponsorInfo, Status as RpcStatus,
+        Log as RpcLog, Receipt as RpcReceipt, RewardInfo as RpcRewardInfo,
+        SendTxRequest, SponsorInfo as RpcSponsorInfo, Status as RpcStatus,
         StorageRoot as RpcStorageRoot, SyncGraphStates,
         Transaction as RpcTransaction, H160 as RpcH160, H256 as RpcH256,
         H520 as RpcH520, U128 as RpcU128, U256 as RpcU256, U64 as RpcU64,
@@ -605,6 +605,7 @@ impl RpcImpl {
                         &mut block_size_limit,
                         num_txs_simple,
                         num_txs_erc20,
+                        &self.consensus.get_config().chain_id,
                     );
 
                 Ok(block_gen.generate_block(
@@ -730,6 +731,29 @@ impl RpcImpl {
             .cloned()
             .map(RpcLog::from)
             .collect())
+    }
+
+    fn get_block_reward_info(
+        &self, epoch: EpochNumber,
+    ) -> RpcResult<Vec<RpcRewardInfo>> {
+        info!(
+            "RPC Request: cfx_getBlockRewardInfo epoch_number={:?}",
+            epoch
+        );
+
+        let blocks = self.consensus.get_block_hashes_by_epoch(epoch.into())?;
+
+        let mut ret = Vec::new();
+        for b in blocks {
+            if let Some(reward_result) = self
+                .consensus
+                .get_data_manager()
+                .block_reward_result_by_hash(&b)
+            {
+                ret.push(RpcRewardInfo::new(b, reward_result));
+            }
+        }
+        Ok(ret)
     }
 
     fn call(
@@ -967,6 +991,7 @@ impl Cfx for CfxHandler {
                 &self, account_addr: RpcH160, contract_addr: RpcH160, gas_limit: RpcU256, gas_price: RpcU256, storage_limit: RpcU256, epoch: Option<EpochNumber>,
             ) -> JsonRpcResult<CheckBalanceAgainstTransactionResponse>;
             fn get_logs(&self, filter: RpcFilter) -> BoxFuture<Vec<RpcLog>>;
+            fn get_block_reward_info(&self, num: EpochNumber) -> JsonRpcResult<Vec<RpcRewardInfo>>;
             fn send_raw_transaction(&self, raw: Bytes) -> JsonRpcResult<RpcH256>;
             fn storage_at(&self, addr: RpcH160, pos: RpcH256, epoch_number: Option<EpochNumber>)
                 -> BoxFuture<Option<RpcH256>>;