+++ client/src/rpc/traits/cfx.rs
@@ -6,7 +6,8 @@ use super::super::types::{
     Account as RpcAccount, Block, Bytes, CallRequest,
     CheckBalanceAgainstTransactionResponse, EpochNumber,
     EstimateGasAndCollateralResponse, Filter as RpcFilter, Log as RpcLog,
-    Receipt as RpcReceipt, SponsorInfo as RpcSponsorInfo, Status as RpcStatus,
+    Receipt as RpcReceipt, RewardInfo as RpcRewardInfo,
+    SponsorInfo as RpcSponsorInfo, Status as RpcStatus,
     StorageRoot as RpcStorageRoot, Transaction, H160 as RpcH160,
     H256 as RpcH256, U256 as RpcU256, U64 as RpcU64,
 };
@@ -207,6 +208,12 @@ pub trait Cfx {
 
     #[rpc(name = "cfx_getStatus")]
     fn get_status(&self) -> JsonRpcResult<RpcStatus>;
+
+    /// Returns block reward information in an epoch
+    #[rpc(name = "cfx_getBlockRewardInfo")]
+    fn get_block_reward_info(
+        &self, num: EpochNumber,
+    ) -> JsonRpcResult<Vec<RpcRewardInfo>>;
 
     //        /// Returns transaction at given block hash and index.
     //        #[rpc(name = "cfx_getTransactionByBlockHashAndIndex")]