+++ network/src/service.rs
@@ -1049,7 +1049,11 @@ impl NetworkServiceInner {
                 let data = sess.readable(io, self);
                 match data {
                     Ok(session_data) => {
-                        token_to_disconnect = session_data.token_to_disconnect;
+                        if session_data.token_to_disconnect.is_some() {
+                            debug!("session_readable: set token_to_disconnect to {:?}", session_data.token_to_disconnect);
+                            token_to_disconnect =
+                                session_data.token_to_disconnect;
+                        }
                         match session_data.session_data {
                             SessionData::Ready => {
                                 handshake_done = true;
@@ -1242,10 +1246,10 @@ impl NetworkServiceInner {
                             );
                         }
                         UpdateNodeOperation::Demotion => {
-                            self.node_db.write().demote(&id);
-                            self.node_db.write().note_failure(
-                                &id, true,  /* by_connection */
-                                false, /* trusted_only */
+                            let mut node_db = self.node_db.write();
+                            node_db.demote(&id);
+                            node_db.note_demoted(
+                                &id, true, /* by_connection */
                             );
                         }
                         UpdateNodeOperation::Remove => {
@@ -1314,10 +1318,10 @@ impl NetworkServiceInner {
                             );
                         }
                         UpdateNodeOperation::Demotion => {
-                            self.node_db.write().demote(node_id);
-                            self.node_db.write().note_failure(
-                                node_id, true,  /* by_connection */
-                                false, /* trusted_only */
+                            let mut node_db = self.node_db.write();
+                            node_db.demote(node_id);
+                            node_db.note_demoted(
+                                node_id, true, /* by_connection */
                             );
                         }
                         UpdateNodeOperation::Remove => {